motor double

+++ src/main/java/bo/motoren/Motor.java
@@ -12,7 +12,7 @@ public abstract class Motor{
 		this.motor = new EV3LargeRegulatedMotor(port);
 	}
 	
-	public void move(int differenz, int time, int umfang) {
+	public void move(double differenz, double time, double umfang) {
 		double gradProSekunde = MathUtil.toDegree(differenz, time, umfang);
 		motor.setSpeed((int)gradProSekunde);
 	}
